% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_Friend.R
\docType{class}
\name{MyAnswer}
\alias{MyAnswer}
\title{My Answer}
\description{
My Answer to a Question
}

\seealso{
Other TheFriend: \code{\link{PoohAnswer}},
  \code{\link{RightAnswer}}, \code{\link{WrongAnswer}}
}
\concept{TheFriend}
