% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\name{One}
\alias{One}
\title{R6 Class Age One}
\description{
Age One
}
\details{
Age One
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-One-new}{\code{One$new()}}
\item \href{#method-One-print_poem}{\code{One$print_poem()}}
\item \href{#method-One-how_old_am_i}{\code{One$how_old_am_i()}}
\item \href{#method-One-clone}{\code{One$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One-new"></a>}}
\if{latex}{\out{\hypertarget{method-One-new}{}}}
\subsection{Method \code{new()}}{
Create a New Age One Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An Age One Object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One-print_poem"></a>}}
\if{latex}{\out{\hypertarget{method-One-print_poem}{}}}
\subsection{Method \code{print_poem()}}{
Print poem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One$print_poem()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One-how_old_am_i"></a>}}
\if{latex}{\out{\hypertarget{method-One-how_old_am_i}{}}}
\subsection{Method \code{how_old_am_i()}}{
Get Age
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One$how_old_am_i()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-One-clone"></a>}}
\if{latex}{\out{\hypertarget{method-One-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{One$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
