% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_Friend.R
\docType{class}
\name{PoohAnswer}
\alias{PoohAnswer}
\title{Pooh's Answer}
\description{
Pooh's Answer to a Question
}

\seealso{
Other TheFriend: \code{\link{MyAnswer}},
  \code{\link{RightAnswer}}, \code{\link{WrongAnswer}}
}
\concept{TheFriend}
