% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\name{Six}
\alias{Six}
\title{R6 Class Age Six}
\description{
Age Six
}
\details{
Age Six
}
\section{Super classes}{
\code{\link[milne:One]{milne::One}} -> \code{\link[milne:Two]{milne::Two}} -> \code{milne::HardlyThree} -> \code{milne::NA} -> \code{\link[milne:Five]{milne::Five}} -> \code{Six}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Six-print_poem}{\code{Six$print_poem()}}
\item \href{#method-Six-clone}{\code{Six$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="milne" data-topic="One" data-id="how_old_am_i"><a href='../../milne/html/One.html#method-One-how_old_am_i'><code>milne::One$how_old_am_i()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milne" data-topic="One" data-id="initialize"><a href='../../milne/html/One.html#method-One-initialize'><code>milne::One$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Six-print_poem"></a>}}
\if{latex}{\out{\hypertarget{method-Six-print_poem}{}}}
\subsection{Method \code{print_poem()}}{
Print poem six times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Six$print_poem()}\if{html}{\out{</div>}}
}

\subsection{Details}{
I should have looked ahead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Six-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Six-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Six$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
