% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\name{Two}
\alias{Two}
\title{R6 Class Age Two}
\description{
Age Two
}
\details{
Age Two
}
\section{Super class}{
\code{\link[milne:One]{milne::One}} -> \code{Two}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Two-print_poem}{\code{Two$print_poem()}}
\item \href{#method-Two-clone}{\code{Two$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="milne" data-topic="One" data-id="how_old_am_i"><a href='../../milne/html/One.html#method-One-how_old_am_i'><code>milne::One$how_old_am_i()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milne" data-topic="One" data-id="initialize"><a href='../../milne/html/One.html#method-One-initialize'><code>milne::One$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Two-print_poem"></a>}}
\if{latex}{\out{\hypertarget{method-Two-print_poem}{}}}
\subsection{Method \code{print_poem()}}{
Print poem two
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Two$print_poem()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Two-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Two-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Two$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
