% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_Friend.R
\docType{class}
\name{WrongAnswer}
\alias{WrongAnswer}
\title{Wrong Answer}
\description{
Incorrect Answer to a Question
}

\seealso{
Other TheFriend: \code{\link{MyAnswer}},
  \code{\link{PoohAnswer}}, \code{\link{RightAnswer}}
}
\concept{TheFriend}
