% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\docType{data}
\name{Five}
\alias{Five}
\title{Age Five}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Five
}
\description{
Age Five
}
\seealso{
Other TheEnd: \code{\link{Four}}, \code{\link{One}},
  \code{\link{Six}}, \code{\link{Three}}, \code{\link{Two}}
}
\concept{TheEnd}
\keyword{datasets}
