% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\docType{data}
\name{Four}
\alias{Four}
\title{Age Four}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Four
}
\description{
Age Four
}
\seealso{
Other TheEnd: \code{\link{Five}}, \code{\link{One}},
  \code{\link{Six}}, \code{\link{Three}}, \code{\link{Two}}
}
\concept{TheEnd}
\keyword{datasets}
