% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\docType{data}
\name{Six}
\alias{Six}
\title{Age Six}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Six
}
\description{
Age Six
}
\seealso{
Other TheEnd: \code{\link{Five}}, \code{\link{Four}},
  \code{\link{One}}, \code{\link{Three}}, \code{\link{Two}}
}
\concept{TheEnd}
\keyword{datasets}
