% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\docType{data}
\name{Three}
\alias{Three}
\title{Age Three}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Three
}
\description{
Age Three
}
\seealso{
Other TheEnd: \code{\link{Five}}, \code{\link{Four}},
  \code{\link{One}}, \code{\link{Six}}, \code{\link{Two}}
}
\concept{TheEnd}
\keyword{datasets}
