% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_End.R
\docType{data}
\name{Two}
\alias{Two}
\title{Age Two}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Two
}
\description{
Age Two
}
\seealso{
Other TheEnd: \code{\link{Five}}, \code{\link{Four}},
  \code{\link{One}}, \code{\link{Six}}, \code{\link{Three}}
}
\concept{TheEnd}
\keyword{datasets}
